LPS-55100 Move InstrumentationAgent to its own package
